requirements.txt
Django>=1.9,<1.10
psycopg2>=2.6
django-compressor>=1.6
django-libsass>=0.6
django-debug-toolbar>=1.4
gunicorn>=19.1.1
Fabric>=1.10.0
django-bootstrap3>=7.0.0
django-money>=0.7.6
django-uuslug>=1.1.7
python-social-auth==0.2.14
sorl-thumbnail>=12.3
South>=1.0.2
